sql-intermediate/tools/sql-tools/pyproject.toml
[project]
name = "sql-tools"
version = "0.1.0"
description = "Herramientas SQL: generador de CSV, importador de CSV a SQL y cargador de datos en bases de datos"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Marco Villarreal" }
]
keywords = ["database", "csv", "postgresql", "sqlserver", "oracle", "data-generation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "Faker>=37.8.0",
    "colorama>=0.4.6",
    "psycopg2-binary>=2.9.9",
    "pyodbc>=5.1.0",
    "oracledb>=2.0.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
csv-generate = "csv_generator.main:main"
csv-import   = "csv_importer.main:main"
db-insert    = "database_inserts.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["csv_generator", "csv_importer", "database_inserts"]
